+++ frontend/app/src/components/chat/index.ts
@@ -1,0 +1,4 @@
+export * from './ChatInput'
+export * from './ChatList'
+export * from './ChatListItem'
+export * from './MemberList'
