+++ frontend/app/src/components/index.ts
@@ -1,2 +1,3 @@
 export * from './utils'
 export * from './profile'
+export * from './chat'
