+++ frontend/app/src/hook/useAuthHeader.ts
@@ -1,4 +1,3 @@
-import axios from 'axios'
 import { useState, useEffect } from 'react'
 
 // TODO: use react-query
