+++ frontend/app/src/view/JoinedRoomList.tsx
@@ -1,4 +1,3 @@
-import { useState, useCallback, useEffect, useRef, useContext } from 'react'
 import { Box, Paper, Stack } from '@mui/material'
 import { styled } from '@mui/material/styles'
 import { JoinedRoom } from 'data'
